fix orders Apartment

+++ src/main/java/com/cg/smart_house/models/Orders.java
@@ -17,7 +17,7 @@ public class Orders implements Serializable,Comparable<Orders>{
     private Long id;
     private Date startTime;
     private Date endTime;
-    private int totalMoney;
+    private Long totalMoney;
 
     @ManyToOne
 //    @JsonIgnore
